Fix captions for the new standard.
The dash delimiter was not working with the package `ccaption` (it was using ':' instead). The `ccaption` package is older and focused on 'continued captions' which the library standard makes no mention to. Updated to the more widely used `caption` package. Checked the output and it matches the one obtained with the docx template.